Arrange the conditinal comp stuff
There are 2 flags
* sdl-resample - will compile the sdl resampler instead of my resampler
* push-audio - will push the audio to the device instead
of the device pulling it by himslef.
