+++ gb/src/audio/wav_file_audio_device.rs
@@ -1,18 +1,14 @@
 use lib_gb::apu::audio_device::*;
+use super::AudioResampler;
 
-#[cfg(not(feature = "sdl-resample"))]
-use crate::audio_resampler::AudioResampler;
-#[cfg(feature = "sdl-resample")]
-use crate::sdl_audio_resampler::SdlAudioResampler as AudioResampler;
-
-pub struct WavfileAudioDevice{
+pub struct WavfileAudioDevice<AR:AudioResampler>{
     target_frequency:u32,
-    resampler: AudioResampler,
+    resampler: AR,
     filename:&'static str,
     samples_buffer:Vec::<StereoSample>
 }
 
-impl WavfileAudioDevice{
+impl<AR:AudioResampler> WavfileAudioDevice<AR>{
     pub fn new(target_freq:u32, original_freq:u32, filename:&'static str)->Self{
         WavfileAudioDevice{
             filename,
@@ -23,13 +19,13 @@ impl WavfileAudioDevice{
     }
 }
 
-impl AudioDevice for WavfileAudioDevice{
+impl<AR:AudioResampler> AudioDevice for WavfileAudioDevice<AR>{
     fn push_buffer(&mut self, buffer:&[StereoSample; BUFFER_SIZE]) {
         self.samples_buffer.append(self.resampler.resample(buffer).as_mut());
     }
 }
 
-impl Drop for WavfileAudioDevice{
+impl<AR:AudioResampler> Drop for WavfileAudioDevice<AR>{
     fn drop(&mut self) {
         let header = wav::header::Header::new(wav::WAV_FORMAT_PCM, 2, self.target_frequency, 16);
         let mut samples = Vec::with_capacity(self.samples_buffer.len() * 2);
